Modify scenario test to persistently check task state cache
This is chosen to make sure the unittest actually testing retried task.
Unittest will use state in the analysis model (the cache) to check if
an analysis is succesful or not. We avoid to check the task result directly,
since it is possible that the task failed but is immediately retried.

In production scenarios, we rely on the immediate task result because
we want to know the immediate status of the task.